requirements.txt
Django
djangorestframework
django-filter
dj-database-url
psycopg2-binary
django-extensions
django-debug-toolbar
ipython
drf-yasg
